requirements.txt
streamlit
plotly

# Data processing
pandas
numpy
pyarrow

# MQTT
paho-mqtt

# Database
psycopg2-binary
sqlalchemy

# Utilities
python-dotenv
python-dateutil
requests
